Added Basic Calibration
Basic Demo Ready Calibration
Note: Sensor 4 Readings are faulty, I will have to recrimp Sensor 4 wires and re-test.

Additionally, There are noise spikes that occur every minute or so on multiple sensors to give an arbitruary idea of a problem to resolve in the future. (I have graphs available in Google Drive showing the spikes)